Inverse Probability Mass Function

Compute the probability mass function for a specified categorical variable conditional on an interval of the continuous random variable's values.

inverse_pmf(x, val.range, var.new, sample.args = list())

Arguments

x
an rvtable.

val.range
range of values of the continuous random variable in the rvtable.

var.new
the categorical variable for which to compute the pmf given val.range.

sample.args
optional arguments used when sampling.

Value

an rvtable.

Details

This function computes the pmf of a categorical variable, providing probabilities corresponding to the levels of the variable, given a specific range of values of the continuous random variable in the rvtable. At this time the rvtable type must be continuous.

Examples

library(data.table)
library(dplyr)
x <- data.table(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)) %>% rvtable
y1 <- inverse_pmf(x, c(5, 8), "id1", sample.args=list(n=5))
y1

#>     Val  id2 id3       Prob
#>  1:   A high   1 0.06666667
#>  2:   A high   2 0.30769231
#>  3:   A  low   1 0.00000000
#>  4:   A  low   2 0.00000000
#>  5:   B high   1 0.20000000
#>  6:   B high   2 0.30769231
#>  7:   B  low   1 0.20000000
#>  8:   B  low   2 0.14285714
#>  9:   C high   1 0.26666667
#> 10:   C high   2 0.15384615
#> 11:   C  low   1 0.30000000
#> 12:   C  low   2 0.28571429
#> 13:   D high   1 0.26666667
#> 14:   D high   2 0.15384615
#> 15:   D  low   1 0.20000000
#> 16:   D  low   2 0.28571429
#> 17:   E high   1 0.20000000
#> 18:   E high   2 0.07692308
#> 19:   E  low   1 0.30000000
#> 20:   E  low   2 0.28571429

x <- filter(x, id2=="low" & id3==1) %>% select(-id2, -id3) %>% rvtable
y2 <- inverse_pmf(x, c(5,8), "id1", sample.args=list(n=5))
y2

#>    Val       Prob
#> 1:   A 0.00000000
#> 2:   B 0.36363636
#> 3:   C 0.18181818
#> 4:   D 0.36363636
#> 5:   E 0.09090909