Merge RV Table Conditional Distributions

Merge conditional distributions of a random variable in an rvtable over levels of ungrouped categorical variables.

merge_rvtable(x, density.args = list(), sample.args = list())

Arguments

x
an rvtable.

density.args
optional arguments passed to density.

sample.args
optional arguments used when sampling.

Value

an rvtable.

Details

Distributions are merged using a cycle of bootstrap resampling followed by density re-estimation. Merging relies simply on what x is grouped by. It assumes equal weights for all levels of all grouping variables over which a random variable is being merged. This is why it is called merging rather than marginalizing. This function is also used by marginalize and can properly obtain a marginal distribution when passed an rvtable including a weights column, which marginalize provides. If calling this function directly, it is important to know that it is generally intended for simple merging. Use marginalize to ensure proper marginal distributions are obtained since it can be used with or without non-constant weights.

Examples

library(data.table)
library(dplyr)
x <- data.table(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)) %>% rvtable
merge_rvtable(x)

#> # A tibble: 512 x 2
#>          Val         Prob
#>        <dbl>        <dbl>
#>  1 0.4203652 6.930976e-05
#>  2 0.4402464 9.686712e-05
#>  3 0.4601275 1.337556e-04
#>  4 0.4800087 1.825274e-04
#>  5 0.4998898 2.462426e-04
#>  6 0.5197710 3.285233e-04
#>  7 0.5396521 4.357114e-04
#>  8 0.5595333 5.756613e-04
#>  9 0.5794145 7.520624e-04
#> 10 0.5992956 9.718283e-04
#> # ... with 502 more rows

x %>% group_by(id1) %>% merge_rvtable

#> # A tibble: 2,560 x 3
#> # Groups:   id1 [5]
#>      id1       Val         Prob
#>    <chr>     <dbl>        <dbl>
#>  1     A 0.1976976 6.941509e-05
#>  2     A 0.2184506 8.961257e-05
#>  3     A 0.2392036 1.147755e-04
#>  4     A 0.2599566 1.458692e-04
#>  5     A 0.2807096 1.839884e-04
#>  6     A 0.3014626 2.303641e-04
#>  7     A 0.3222156 2.883394e-04
#>  8     A 0.3429686 3.596784e-04
#>  9     A 0.3637216 4.454355e-04
#> 10     A 0.3844746 5.477703e-04
#> # ... with 2,550 more rows

x %>% group_by(id1, id2) %>% merge_rvtable

#> # A tibble: 5,120 x 4
#> # Groups:   id1, id2 [10]
#>      id1    id2        Val         Prob
#>    <chr> <fctr>      <dbl>        <dbl>
#>  1     A   high 0.08223904 6.919104e-05
#>  2     A   high 0.10344464 8.639328e-05
#>  3     A   high 0.12465024 1.080567e-04
#>  4     A   high 0.14585584 1.342608e-04
#>  5     A   high 0.16706144 1.657400e-04
#>  6     A   high 0.18826705 2.033048e-04
#>  7     A   high 0.20947265 2.478443e-04
#>  8     A   high 0.23067825 3.030136e-04
#>  9     A   high 0.25188385 3.685100e-04
#> 10     A   high 0.27308945 4.454803e-04
#> # ... with 5,110 more rows