Random Variable Table

Class constructor for rvtable objects.

rvtable(x, y = NULL, Val = "Val", Prob = "Prob", discrete = FALSE,
  density.args = list())

Arguments

x
a numeric vector, data frame, or data table.

y
an optional vector of probabilities associated with x when x is a numeric vector with no similar probabilities vector attribute.

Val
the column name of x referring to random variable values when x is a data frame or data table.

Prob
the column name of x referring to random variable values when x is a data frame or data table.

discrete
whether the random variable is discrete.

density.args
optional arguments passed to density.

Value

an S3 object of class rvtable

Details

These are long format data tables containing a Val and Prob column describing the distribution of a random variable. Any other columns are ID columns and should be categorical variables. The random variable described by Val and Prob may be discrete or continuous. When discrete, probabilities are true probabilities. When continuous, Val and Prob are based on x and y output from density and describe a distribution curve, and therefore values in Prob may be greater than one and may not sum to one. Val is typically numeric but may be character when discrete such as when an rvtable object is returned from inverse_pmf. All rvtable objects are either distribution-based or sample-based. This primary constructor only constructs distribution- or sample-based rvtable objects, with the attribute tabletype="distribution" or tabletype="sample". Sampling on an rvtable can generate a sample-based rvtable, with the attribute tabletype="sample". Every rvtable object also has a variable type attribute, rvtype, which is either "discrete" or "continuous".

Examples

# basic samples from continuous and discrete RVs
x <- rnorm(100)
rvtable(x)

#>            Val         Prob
#>   1: -3.454880 0.0002131446
#>   2: -3.438980 0.0002396017
#>   3: -3.423080 0.0002688230
#>   4: -3.407179 0.0003023149
#>   5: -3.391279 0.0003388508
#>  ---                       
#> 508:  4.606558 0.0001635291
#> 509:  4.622458 0.0001469351
#> 510:  4.638358 0.0001316073
#> 511:  4.654258 0.0001180818
#> 512:  4.670159 0.0001057418

rvtable(x, density.args=list(n=50, adjust=2))

#>             Val         Prob
#>  1: -4.73353012 1.913517e-04
#>  2: -4.51552321 4.282042e-04
#>  3: -4.29751629 9.063360e-04
#>  4: -4.07950938 1.811232e-03
#>  5: -3.86150247 3.421668e-03
#>  6: -3.64349556 6.136505e-03
#>  7: -3.42548865 1.043900e-02
#>  8: -3.20748174 1.689552e-02
#>  9: -2.98947483 2.608865e-02
#> 10: -2.77146792 3.849002e-02
#> 11: -2.55346101 5.442074e-02
#> 12: -2.33545410 7.389227e-02
#> 13: -2.11744719 9.653380e-02
#> 14: -1.89944028 1.215849e-01
#> 15: -1.68143337 1.478483e-01
#> 16: -1.46342646 1.738532e-01
#> 17: -1.24541955 1.980107e-01
#> 18: -1.02741264 2.189284e-01
#> 19: -0.80940573 2.356044e-01
#> 20: -0.59139882 2.475857e-01
#> 21: -0.37339191 2.549854e-01
#> 22: -0.15538500 2.581914e-01
#> 23:  0.06262191 2.576780e-01
#> 24:  0.28062882 2.537518e-01
#> 25:  0.49863573 2.464408e-01
#> 26:  0.71664264 2.356778e-01
#> 27:  0.93464955 2.214267e-01
#> 28:  1.15265646 2.038732e-01
#> 29:  1.37066337 1.835915e-01
#> 30:  1.58867028 1.614397e-01
#> 31:  1.80667719 1.384907e-01
#> 32:  2.02468410 1.158589e-01
#> 33:  2.24269101 9.455010e-02
#> 34:  2.46069792 7.533454e-02
#> 35:  2.67870483 5.868086e-02
#> 36:  2.89671174 4.478701e-02
#> 37:  3.11471865 3.354767e-02
#> 38:  3.33272556 2.469774e-02
#> 39:  3.55073247 1.788525e-02
#> 40:  3.76873938 1.271563e-02
#> 41:  3.98674629 8.851253e-03
#> 42:  4.20475320 6.006452e-03
#> 43:  4.42276011 3.949674e-03
#> 44:  4.64076702 2.505003e-03
#> 45:  4.85877393 1.523635e-03
#> 46:  5.07678084 8.839484e-04
#> 47:  5.29478775 4.883929e-04
#> 48:  5.51279467 2.555986e-04
#> 49:  5.73080158 1.264016e-04
#> 50:  5.94880849 5.910205e-05
#>             Val         Prob

rvtable(x, discrete=TRUE) # incorrect: x is a continuous RV

#>              Val Prob
#>   1: -2.17623045 0.01
#>   2: -2.11969978 0.01
#>   3: -1.86738529 0.01
#>   4: -1.86592370 0.01
#>   5: -1.85088977 0.01
#>   6: -1.80025993 0.01
#>   7: -1.78106193 0.01
#>   8: -1.64290624 0.01
#>   9: -1.46864337 0.01
#>  10: -1.42565951 0.01
#>  11: -1.37076122 0.01
#>  12: -1.31633764 0.01
#>  13: -1.27781992 0.01
#>  14: -1.24619198 0.01
#>  15: -1.14312848 0.01
#>  16: -1.08900955 0.01
#>  17: -0.97967535 0.01
#>  18: -0.96767371 0.01
#>  19: -0.95260348 0.01
#>  20: -0.94410166 0.01
#>  21: -0.92066340 0.01
#>  22: -0.86072428 0.01
#>  23: -0.82776894 0.01
#>  24: -0.80534353 0.01
#>  25: -0.78938814 0.01
#>  26: -0.78447669 0.01
#>  27: -0.78207767 0.01
#>  28: -0.77928540 0.01
#>  29: -0.77705556 0.01
#>  30: -0.77219209 0.01
#>  31: -0.71605867 0.01
#>  32: -0.71128732 0.01
#>  33: -0.70859443 0.01
#>  34: -0.69120544 0.01
#>  35: -0.68326687 0.01
#>  36: -0.67909543 0.01
#>  37: -0.66756477 0.01
#>  38: -0.63969296 0.01
#>  39: -0.46251911 0.01
#>  40: -0.43219524 0.01
#>  41: -0.32849189 0.01
#>  42: -0.28362275 0.01
#>  43: -0.28045548 0.01
#>  44: -0.23854626 0.01
#>  45: -0.23564615 0.01
#>  46: -0.23276176 0.01
#>  47: -0.21714460 0.01
#>  48: -0.20889688 0.01
#>  49:  0.04433469 0.01
#>  50:  0.11759261 0.01
#>  51:  0.13191018 0.01
#>  52:  0.19439244 0.01
#>  53:  0.20538939 0.01
#>  54:  0.26726762 0.01
#>  55:  0.29738725 0.01
#>  56:  0.35227917 0.01
#>  57:  0.38624909 0.01
#>  58:  0.40152349 0.01
#>  59:  0.40427794 0.01
#>  60:  0.42123042 0.01
#>  61:  0.46205231 0.01
#>  62:  0.49790884 0.01
#>  63:  0.57991478 0.01
#>  64:  0.58807712 0.01
#>  65:  0.63747287 0.01
#>  66:  0.67638957 0.01
#>  67:  0.69362435 0.01
#>  68:  0.70005531 0.01
#>  69:  0.70359786 0.01
#>  70:  0.71540139 0.01
#>  71:  0.71752478 0.01
#>  72:  0.73762945 0.01
#>  73:  0.74786883 0.01
#>  74:  0.77599004 0.01
#>  75:  0.79245114 0.01
#>  76:  0.91107847 0.01
#>  77:  0.91187392 0.01
#>  78:  0.94433756 0.01
#>  79:  1.05542231 0.01
#>  80:  1.13891313 0.01
#>  81:  1.15208976 0.01
#>  82:  1.20402895 0.01
#>  83:  1.21450968 0.01
#>  84:  1.33985992 0.01
#>  85:  1.38950587 0.01
#>  86:  1.45054321 0.01
#>  87:  1.53944942 0.01
#>  88:  1.58499677 0.01
#>  89:  1.59617200 0.01
#>  90:  1.60994296 0.01
#>  91:  1.61938957 0.01
#>  92:  1.74471328 0.01
#>  93:  1.81503083 0.01
#>  94:  1.86113580 0.01
#>  95:  1.95716996 0.01
#>  96:  2.04300891 0.01
#>  97:  2.13270193 0.01
#>  98:  2.58443224 0.01
#>  99:  2.73610845 0.01
#> 100:  3.39150882 0.01
#>              Val Prob

x <- sample(1:10, size=30, replace=TRUE, prob=sqrt(10:1))
rvtable(x, discrete=TRUE) # discrete=T only needed if ambiguous

#>    Val       Prob
#> 1:   1 0.16666667
#> 2:   2 0.10000000
#> 3:   3 0.13333333
#> 4:   4 0.03333333
#> 5:   5 0.16666667
#> 6:   6 0.13333333
#> 7:   7 0.13333333
#> 8:   8 0.06666667
#> 9:   9 0.06666667

x <- 1:5
probs <- c(0.1, 0.2, 0.3, 0.15, 0.25)
rvtable(x, y=probs) # discrete inferred from y

#>    Val Prob
#> 1:   1 0.10
#> 2:   2 0.20
#> 3:   3 0.30
#> 4:   4 0.15
#> 5:   5 0.25

attr(x, "probabilities") <- probs
rvtable(x) # discrete inferred from attributes

#>    Val Prob
#> 1:   1 0.10
#> 2:   2 0.20
#> 3:   3 0.30
#> 4:   4 0.15
#> 5:   5 0.25

# an existing data frame or data table
x <- data.frame(Val=1:10, Prob=0.1)
rvtable(x)

#>     Val Prob
#>  1:   1  0.1
#>  2:   2  0.1
#>  3:   3  0.1
#>  4:   4  0.1
#>  5:   5  0.1
#>  6:   6  0.1
#>  7:   7  0.1
#>  8:   8  0.1
#>  9:   9  0.1
#> 10:  10  0.1

library(data.table)
x <- data.table(id=rep(LETTERS[1:2], each=10), v1=rep(1:10, 2), p1=c(rep(0.1, 10), sqrt(1:10)))
rvtable(x, Val="v1", Prob="p1")

#>     id Val     Prob
#>  1:  A   1 0.100000
#>  2:  A   2 0.100000
#>  3:  A   3 0.100000
#>  4:  A   4 0.100000
#>  5:  A   5 0.100000
#>  6:  A   6 0.100000
#>  7:  A   7 0.100000
#>  8:  A   8 0.100000
#>  9:  A   9 0.100000
#> 10:  A  10 0.100000
#> 11:  B   1 1.000000
#> 12:  B   2 1.414214
#> 13:  B   3 1.732051
#> 14:  B   4 2.000000
#> 15:  B   5 2.236068
#> 16:  B   6 2.449490
#> 17:  B   7 2.645751
#> 18:  B   8 2.828427
#> 19:  B   9 3.000000
#> 20:  B  10 3.162278